Plot Ceteris Paribus Explanations

Function 'plot.ceteris_paribus_explainer' plots Ceteris Paribus Plots for selected observations. Various parameters help to decide what should be plotted, profiles, aggregated profiles, points or rugs.

# S3 method for ceteris_paribus_explainer
plot(x, ..., size = 1, alpha = 0.3,
  color = "black", size_points = 2, alpha_points = 1,
  color_points = color, size_rugs = 0.5, alpha_rugs = 1,
  color_rugs = color, size_residuals = 1, alpha_residuals = 1,
  color_residuals = color, only_numerical = TRUE, show_profiles = TRUE,
  show_observations = TRUE, show_rugs = FALSE, show_residuals = FALSE,
  aggregate_profiles = NULL, facet_ncol = NULL, selected_variables = NULL)

Arguments

x
a ceteris paribus explainer produced with function `ceteris_paribus()`

...
other explainers that shall be plotted together

size
a numeric. Size of lines to be plotted

alpha
a numeric between 0 and 1. Opacity of lines

color
a character. Either name of a color or name of a variable that should be used for coloring

size_points
a numeric. Size of points to be plotted

alpha_points
a numeric between 0 and 1. Opacity of points

color_points
a character. Either name of a color or name of a variable that should be used for coloring

size_rugs
a numeric. Size of rugs to be plotted

alpha_rugs
a numeric between 0 and 1. Opacity of rugs

color_rugs
a character. Either name of a color or name of a variable that should be used for coloring

size_residuals
a numeric. Size of line and points to be plotted for residuals

alpha_residuals
a numeric between 0 and 1. Opacity of points and lines for residuals

color_residuals
a character. Either name of a color or name of a variable that should be used for coloring for residuals

only_numerical
a logical. If TRUE then only numerical variables will be plotted. If FALSE then only categorical variables will be plotted.

show_profiles
a logical. If TRUE then profiles will be plotted. Either individual or aggregate (see `aggregate_profiles`)

show_observations
a logical. If TRUE then individual observations will be marked as points

show_rugs
a logical. If TRUE then individual observations will be marked as rugs

show_residuals
a logical. If TRUE then residuals will be plotted as a line ended with a point

aggregate_profiles
function. If NULL (default) then individual profiles will be plotted. If a function (e.g. mean or median) then profiles will be aggregated and only the aggregate profile will be plotted

facet_ncol
number of columns for the `facet_wrap()`

selected_variables
if not NULL then only `selected_variables` will be presented

Value

a ggplot2 object

Examples

library("DALEX")

library("randomForest")
set.seed(59)

apartments_rf <- randomForest(m2.price ~ construction.year + surface + floor +
                                no.rooms + district, data = apartments)
explainer_rf <- explain(apartments_rf,
                        data = apartmentsTest[,2:6], y = apartmentsTest$m2.price)

apartments_lm <- lm(m2.price ~ construction.year + surface + floor +
                                no.rooms + district, data = apartments)
explainer_lm <- explain(apartments_lm,
                        data = apartmentsTest[,2:6], y = apartmentsTest$m2.price)

library("e1071")
apartments_svm <- svm(m2.price ~ construction.year + surface + floor +
                                no.rooms + district, data = apartments)
explainer_svm <- explain(apartments_svm,
                        data = apartmentsTest[,2:6], y = apartmentsTest$m2.price)

# individual explanations
my_apartment <- apartmentsTest[1, ]

# for random forest
lp_rf <- local_profile(explainer_rf, my_apartment)

#> Error in local_profile.default(model, data, predict_function, new_observation = new_observation,     label = label, variables = variables, grid_points = grid_points,     y = y, ...): could not find function "local_profile.default"

lp_rf

#> Error in eval(expr, envir, enclos): object 'lp_rf' not found

plot(lp_rf)

#> Error in plot(lp_rf): object 'lp_rf' not found

# for others
lp_lm <- local_profile(explainer_lm, my_apartment)

#> Error in local_profile.default(model, data, predict_function, new_observation = new_observation,     label = label, variables = variables, grid_points = grid_points,     y = y, ...): could not find function "local_profile.default"

plot(lp_rf, lp_lm, color = "_label_")

#> Error in plot(lp_rf, lp_lm, color = "_label_"): object 'lp_rf' not found

# for others
lp_svm <- local_profile(explainer_svm, my_apartment)

#> Error in local_profile.default(model, data, predict_function, new_observation = new_observation,     label = label, variables = variables, grid_points = grid_points,     y = y, ...): could not find function "local_profile.default"

plot(lp_rf, lp_lm, lp_svm, color = "_label_")

#> Error in plot(lp_rf, lp_lm, lp_svm, color = "_label_"): object 'lp_rf' not found

# more parameters
plot(lp_rf, lp_lm, lp_svm, color = "_label_",
   show_profiles = TRUE, show_observations = TRUE,
   show_rugs = TRUE,
   alpha = 0.3, alpha_points = 1,
   size_points = 5, size_rugs = 0.5)

#> Error in plot(lp_rf, lp_lm, lp_svm, color = "_label_", show_profiles = TRUE,     show_observations = TRUE, show_rugs = TRUE, alpha = 0.3,     alpha_points = 1, size_points = 5, size_rugs = 0.5): object 'lp_rf' not found

plot(lp_rf, show_profiles = TRUE, show_observations = TRUE,
   color = "black",
   selected_variables = c("construction.year", "surface"),
   alpha = 0.3, alpha_points = 1, alpha_residuals = 0.5,
   size_points = 5, size_rugs = 0.5)

#> Error in plot(lp_rf, show_profiles = TRUE, show_observations = TRUE, color = "black",     selected_variables = c("construction.year", "surface"), alpha = 0.3,     alpha_points = 1, alpha_residuals = 0.5, size_points = 5,     size_rugs = 0.5): object 'lp_rf' not found

plot(lp_rf, show_profiles = TRUE, show_observations = TRUE,
   color = "surface",
   selected_variables = c("construction.year", "surface"),
   alpha = 0.3, alpha_points = 1, alpha_residuals = 0.5,
   size_points = 5, size_rugs = 0.5)

#> Error in plot(lp_rf, show_profiles = TRUE, show_observations = TRUE, color = "surface",     selected_variables = c("construction.year", "surface"), alpha = 0.3,     alpha_points = 1, alpha_residuals = 0.5, size_points = 5,     size_rugs = 0.5): object 'lp_rf' not found

plot(lp_rf, show_profiles = TRUE, show_observations = TRUE,
   color = "darkblue", aggregate_profiles = mean,
   selected_variables = c("construction.year", "surface"),
   alpha = 0.3, alpha_points = 1, alpha_residuals = 0.5,
   size_points = 5, size_rugs = 0.5)

#> Error in plot(lp_rf, show_profiles = TRUE, show_observations = TRUE, color = "darkblue",     aggregate_profiles = mean, selected_variables = c("construction.year",         "surface"), alpha = 0.3, alpha_points = 1, alpha_residuals = 0.5,     size_points = 5, size_rugs = 0.5): object 'lp_rf' not found

# --------
# multiclass

HR_rf <- randomForest(status ~ . , data = HR)
explainer_rf <- explain(HR_rf,
                        data = HRTest,
                        y = HRTest)

my_HR <- HRTest[1, ]

lp_rf <- local_profile(explainer_rf,
                       my_HR)

#> Error in local_profile.default(model, data, predict_function, new_observation = new_observation,     label = label, variables = variables, grid_points = grid_points,     y = y, ...): could not find function "local_profile.default"

lp_rf

#> Error in eval(expr, envir, enclos): object 'lp_rf' not found

plot(lp_rf, color = "_label_")

#> Error in plot(lp_rf, color = "_label_"): object 'lp_rf' not found